+++ lib/src/animations/animations.dart
@@ -1,3 +1,1 @@
-export 'calendar_height_animation.dart';
-export 'calendar_scale_animation.dart';
-export 'reverse/reverse.dart';
+export 'calendar_animations.dart';