+++ lib/src/config/package_config.dart
@@ -13,7 +13,9 @@ const double calendarWeekdaysHorizontalPadding = 12.0;
 const double calendarWeekdaysHeight = 18.0;
 const double calendarWidth = 320.0;
 const double calendarHeight = 332.0;
-const double calendarMaxHeight = 335.0;
+const double calendarBlurAmount = 40.0;
+const int calendarBluredLightBackgroundColorAlpha = 180;
+const int calendarBluredDarkBackgroundColorAlpha = 160;
 
 // ANIMATION
 
@@ -24,3 +26,8 @@ const Cubic calendarAnimationCurve = Curves.ease;
 const Duration calendarHeaderFadeDuration = Duration(milliseconds: 300);
 const Duration calendarYearPickerFadeDuration = Duration(milliseconds: 250);
 const Duration calendarYearPickerDuration = Duration(milliseconds: 500);
+
+// Other
+
+const String calendarPickerRouteName = 'CupertinoCalendarPicker';
+const String calendarPickerBarrierLabel = 'CupertinoCalendarPickerBarrier';